+++ sdk/__tests__/port/in/Security.test.ts
@@ -1387,7 +1387,7 @@ describe('🧪 Security tests', () => {
     const vcBase64 = await createVcT3();
     const decodedVC = Buffer.from(vcBase64, 'base64').toString('utf-8');
     const vcJson = JSON.parse(decodedVC);
-    const oneSecondBeforeNow = new Date(Date.now() - 1000).toISOString(); // 1 second before now
+    const oneSecondBeforeNow = new Date(Date.now() - 1000).toISOString();
     vcJson.validUntil = oneSecondBeforeNow;
     const corruptedVcJson = JSON.stringify(vcJson);
     const wrongVcBase64 = Buffer.from(corruptedVcJson).toString('base64');
