+++ sdk/src/app/usecase/command/security/operations/kyc/grantKyc/GrantKYCCommandHandler.ts
@@ -218,6 +218,10 @@ import { RPCQueryAdapter } from '../../../../../../../port/out/rpc/RPCQueryAdapt
 import { SecurityPaused } from '../../../error/SecurityPaused.js';
 import { SecurityRole } from '../../../../../../../domain/context/security/SecurityRole.js';
 import { NotGrantedRole } from '../../../error/NotGrantedRole.js';
+import { Terminal3VC } from '../../../../../../../domain/context/kyc/terminal3.js';
+import { verifyVc } from '@terminal3/verify_vc';
+import { SignedCredential } from '@terminal3/vc_core';
+import { MaxUint256 } from '@ethersproject/constants';
 
 @CommandHandler(GrantKYCCommand)
 export class GrantKYCCommandHandler
@@ -237,7 +241,24 @@ export class GrantKYCCommandHandler
   ) {}
 
   async execute(command: GrantKYCCommand): Promise<GrantKYCCommandResponse> {
-    const { securityId, targetId, VCId, issuer, validFrom, validTo } = command;
+    const { securityId, targetId, vcBase64 } = command;
+
+    const signedCredential: SignedCredential =
+      Terminal3VC.vcFromBase64(vcBase64);
+    const verificationResult = await verifyVc(signedCredential);
+    if (!verificationResult.isValid) {
+      throw new Error('Invalid VC');
+    }
+
+    const issuer: string = signedCredential.issuer.split(':').pop()!;
+
+    signedCredential.validFrom = !signedCredential.validFrom
+      ? Date.parse(new Date().toString()).toString()
+      : Date.parse(signedCredential.validFrom).toString();
+
+    signedCredential.validUntil = !signedCredential.validUntil
+      ? MaxUint256.toString()
+      : Date.parse(signedCredential.validUntil).toString();
 
     const handler = this.transactionService.getHandler();
     const account = this.accountService.getCurrentAccount();
@@ -279,8 +300,8 @@ export class GrantKYCCommandHandler
       : new EvmAddress(targetId);
 
     if (
-      BigDecimal.fromString(validFrom).isGreaterThan(
-        BigDecimal.fromString(validTo),
+      BigDecimal.fromString(signedCredential.validFrom).isGreaterThan(
+        BigDecimal.fromString(signedCredential.validUntil),
       )
     ) {
       throw new Error('validTo must be later than validFrom.');
@@ -289,9 +310,9 @@ export class GrantKYCCommandHandler
     const res = await handler.grantKYC(
       securityEvmAddress,
       targetEvmAddress,
-      VCId,
-      BigDecimal.fromString(validFrom),
-      BigDecimal.fromString(validTo),
+      signedCredential.id,
+      BigDecimal.fromString(signedCredential.validFrom),
+      BigDecimal.fromString(signedCredential.validUntil),
       issuerEvmAddress,
     );
